+++ src/codewars/ArrayDif.py
