+++ src/codewars/ExtractDomain.py
