+++ src/codewars/LongestConsec.java
@@ -1,10 +1,11 @@
+package codewars;
+
 /**
  * 	Difficulty: 6 Kyu
  *  You are given an array 'strarr' of strings and an integer 'k'. 
  *  Your task is to return the first longest string consisting of k consecutive strings taken in the array.
- * 	@author Nick Pierre
- *
  */
+
 public class LongestConsec {	    
 	public static String longestConsec(String[] strarr, int k) {
