+++ src/codewars/Multiples.py
@@ -3,8 +3,6 @@
 # If we list all the natural numbers below 10 that are multiples of 3 or 5, we get 3, 5, 6 and 9. The sum of these multiples is 23.
 # Finish the solution so that it returns the sum of all the multiples of 3 or 5 below the number passed in.
 #
-# @author Nick Pierre
-
 
 def solution(number):
     sum = 0
@@ -12,4 +10,3 @@ def solution(number):
         if i % 3 == 0 or i % 5 == 0:
             sum += i
     return sum
-
