+++ src/codewars/isPrime.py
